requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
anthropic==0.26.1
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appnope==0.1.4
arxiv==2.1.0
asttokens==2.4.1
attrs==23.2.0
autopep8==2.1.1
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
cryptography==42.0.7
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.12.1
dataclasses-json==0.6.6
debugpy==1.8.1
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
demogpt==1.2.6.33
Deprecated==1.2.14
diff-match-patch==20230430
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
duckduckgo_search==6.1.1
effdet==0.4.1
elastic-transport==8.13.1
elasticsearch==8.13.2
email_validator==2.1.1
emoji==2.12.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.14.0
filetype==1.2.0
flake8==7.0.0
flatbuffers==24.3.25
fonttools==4.52.3
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-vision==3.7.2
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
greenlet==3.0.3
grep-ast==0.3.2
groq==0.8.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
html5lib==1.1
htmldocx==0.0.6
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
inflection==0.5.1
iopath==0.1.10
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.4.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.2.1
langchain-anthropic==0.1.13
langchain-community==0.2.1
langchain-core==0.2.1
langchain-experimental==0.0.59
langchain-google-genai==1.0.5
langchain-groq==0.1.4
langchain-huggingface==0.0.1
langchain-mistralai==0.1.7
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langchain-together==0.1.2
langdetect==1.0.9
langsmith==0.1.63
layoutparser==0.3.4
litellm==1.37.16
llama-index==0.10.40
llama-index-agent-openai==0.2.5
llama-index-cli==0.1.12
llama-index-core==0.10.40
llama-index-embeddings-ollama==0.1.2
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-ollama==0.1.5
llama-index-llms-openai==0.1.21
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.23
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-elasticsearch==0.2.0
llama-parse==0.4.4
llamaindex-py-client==0.1.19
lxml==5.2.2
lxml_html_clean==0.1.1
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.7.0
md2pdf==1.0.1
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
numexpr==2.10.0
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.0
openai==1.30.1
opencv-python==4.9.0.80
openpyxl==3.1.3
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-anthropic==0.21.1
opentelemetry-instrumentation-bedrock==0.21.1
opentelemetry-instrumentation-chromadb==0.21.1
opentelemetry-instrumentation-cohere==0.21.1
opentelemetry-instrumentation-haystack==0.21.1
opentelemetry-instrumentation-langchain==0.21.1
opentelemetry-instrumentation-llamaindex==0.21.1
opentelemetry-instrumentation-milvus==0.21.1
opentelemetry-instrumentation-mistralai==0.21.1
opentelemetry-instrumentation-ollama==0.21.1
opentelemetry-instrumentation-openai==0.21.1
opentelemetry-instrumentation-pinecone==0.21.1
opentelemetry-instrumentation-qdrant==0.21.1
opentelemetry-instrumentation-replicate==0.21.1
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
opentelemetry-instrumentation-transformers==0.21.1
opentelemetry-instrumentation-urllib3==0.45b0
opentelemetry-instrumentation-vertexai==0.21.1
opentelemetry-instrumentation-watsonx==0.21.1
opentelemetry-instrumentation-weaviate==0.21.1
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-semantic-conventions-ai==0.2.0
opentelemetry-util-http==0.45b0
ordered-set==4.1.0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.9.0
pikepdf==9.0.0
pillow==10.3.0
pillow_heif==0.16.0
platformdirs==4.2.2
playwright==1.44.0
portalocker==2.8.2
posthog==3.5.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pydyf==0.10.0
pyee==11.1.0
pyflakes==3.2.0
Pygments==2.18.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.30.0
pyphen==0.15.0
pyreqwest_impersonate==0.4.5
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-file==2.1.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.0
sentence-transformers==2.7.0
sentry-sdk==2.3.1
setuptools==70.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
streamlit==1.34.0
striprtf==0.0.26
sympy==1.12
tabulate==0.9.0
tavily-python==0.3.3
tenacity==8.3.0
text-generation==0.7.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.3
tinycss2==1.3.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traceloop-sdk==0.21.1
traitlets==5.14.3
transformers==4.41.1
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
unstructured==0.14.2
unstructured-client==0.22.0
unstructured-inference==0.7.33
unstructured.pytesseract==0.3.12
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
weasyprint==62.1
webencodings==0.5.1
websockets==12.0
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.18.2
zopfli==0.2.3
